backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
sendgrid==6.11.0
reportlab==4.0.7
fpdf2==2.7.6
email-validator==2.1.0
requests==2.31.0
